requirements.txt
# 核心 Web 框架
Flask>=2.0.0
Pillow>=9.0.0
gunicorn>=20.1.0